+++ src/library/renderer/RenderTileMap.cpp
@@ -6,7 +6,7 @@
 
 #include "Logger.h"
 namespace PR {
-RenderTileMap::RenderTileMap(size_t xcount, size_t ycount, size_t tilewidth, size_t tileheight)
+RenderTileMap::RenderTileMap(uint32 xcount, uint32 ycount, uint32 tilewidth, uint32 tileheight)
 	: mTileXCount(xcount)
 	, mTileYCount(ycount)
 	, mTileWidth(tilewidth)
@@ -38,10 +38,10 @@ void RenderTileMap::init(const RenderContext& context, TileMode mode)
 	switch (mode) {
 	default:
 	case TM_LINEAR:
-		for (size_t i = 0; i < mTileYCount; ++i) {
-			for (size_t j = 0; j < mTileXCount; ++j) {
-				size_t sx = j * mTileWidth;
-				size_t sy = i * mTileHeight;
+		for (uint32 i = 0; i < mTileYCount; ++i) {
+			for (uint32 j = 0; j < mTileXCount; ++j) {
+				uint32 sx = j * mTileWidth;
+				uint32 sy = i * mTileHeight;
 
 				if (sx < context.width() - 1
 					&& sy < context.height() - 1) {
@@ -58,12 +58,12 @@ void RenderTileMap::init(const RenderContext& context, TileMode mode)
 		}
 		break;
 	case TM_TILE: {
-		size_t k = 0;
+		uint32 k = 0;
 		// Even
-		for (size_t i = 0; i < mTileYCount; ++i) {
-			for (size_t j = ((i % 2) ? 1 : 0); j < mTileXCount; j += 2) {
-				size_t sx = j * mTileWidth;
-				size_t sy = i * mTileHeight;
+		for (uint32 i = 0; i < mTileYCount; ++i) {
+			for (uint32 j = ((i % 2) ? 1 : 0); j < mTileXCount; j += 2) {
+				uint32 sx = j * mTileWidth;
+				uint32 sy = i * mTileHeight;
 
 				if (sx < context.width() - 1
 					&& sy < context.height() - 1) {
@@ -80,10 +80,10 @@ void RenderTileMap::init(const RenderContext& context, TileMode mode)
 			}
 		}
 		// Odd
-		for (size_t i = 0; i < mTileYCount; ++i) {
-			for (size_t j = ((i % 2) ? 0 : 1); j < mTileXCount; j += 2) {
-				size_t sx = j * mTileWidth;
-				size_t sy = i * mTileHeight;
+		for (uint32 i = 0; i < mTileYCount; ++i) {
+			for (uint32 j = ((i % 2) ? 0 : 1); j < mTileXCount; j += 2) {
+				uint32 sx = j * mTileWidth;
+				uint32 sy = i * mTileHeight;
 
 				if (sx < context.width() - 1
 					&& sy < context.height() - 1) {
@@ -102,7 +102,7 @@ void RenderTileMap::init(const RenderContext& context, TileMode mode)
 	} break;
 	case TM_SPIRAL: {
 		MinRadiusGenerator<2> generator(std::max(mTileXCount / 2, mTileYCount / 2));
-		size_t i = 0;
+		uint32 i = 0;
 		while (generator.hasNext()) {
 			const auto p  = generator.next();
 			const auto tx = mTileXCount / 2 + p[0];
@@ -129,8 +129,8 @@ void RenderTileMap::init(const RenderContext& context, TileMode mode)
 
 RenderTile* RenderTileMap::getNextTile(uint32 maxSample)
 {
-	for (size_t i = 0; i < mTileYCount; ++i) {
-		for (size_t j = 0; j < mTileXCount; ++j) {
+	for (uint32 i = 0; i < mTileYCount; ++i) {
+		for (uint32 j = 0; j < mTileXCount; ++j) {
 			// TODO: Better check up for AS
 			if (mTileMap[i * mTileXCount + j]
 				&& mTileMap[i * mTileXCount + j]->samplesRendered() < maxSample
@@ -148,8 +148,8 @@ RenderTile* RenderTileMap::getNextTile(uint32 maxSample)
 RenderTileStatistics RenderTileMap::statistics() const
 {
 	RenderTileStatistics s;
-	for (size_t i = 0; i < mTileYCount; ++i) {
-		for (size_t j = 0; j < mTileXCount; ++j) {
+	for (uint32 i = 0; i < mTileYCount; ++i) {
+		for (uint32 j = 0; j < mTileXCount; ++j) {
 			s += mTileMap[i * mTileXCount + j]->statistics();
 		}
 	}
@@ -158,8 +158,8 @@ RenderTileStatistics RenderTileMap::statistics() const
 
 void RenderTileMap::reset()
 {
-	for (size_t i = 0; i < mTileYCount; ++i)
-		for (size_t j = 0; j < mTileXCount; ++j)
+	for (uint32 i = 0; i < mTileYCount; ++i)
+		for (uint32 j = 0; j < mTileXCount; ++j)
 			mTileMap[i * mTileXCount + j]->reset();
 }
 } // namespace PR
